Switch more funcs to long double, and return results as NV

+++ util.h
@@ -18,8 +18,8 @@ extern UV  _XS_nth_prime(UV x);
 extern signed char* _moebius_range(UV low, UV high);
 extern UV*    _totient_range(UV low, UV high);
 extern IV     mertens(UV n);
-extern double _XS_chebyshev_theta(UV n);
-extern double _XS_chebyshev_psi(UV n);
+extern long double _XS_chebyshev_theta(UV n);
+extern long double _XS_chebyshev_psi(UV n);
 
 extern long double _XS_ExponentialIntegral(long double x);
 extern long double _XS_LogarithmicIntegral(long double x);
